+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Ideas/Upcoming]
 
 ## [Unreleased]
+### Changed
+* Replaced KeyboardWatcher by InputWatcher
 
 ## [1.1.3] - 2022-05-14
 ### Added